Extract osx-specific tests in their own file
Includes making the spec helpers vary according to the platform
